Extract parent test case class MockSWFTestCase to handle setUp/tearDown better
When trying to add a second mocked test, it appeared that moto.swf state
was not reset, so many calls were failing. Extracting those steps in a
parent class will help getting the correct behaviour and reduce
boilerplate.
