+++ tests/utils/__init__.py
@@ -1,0 +1,2 @@
+from .integration_test_case import IntegrationTestCase  # noqa
+from .mock_swf_test_case import MockSWFTestCase  # noqa
