+++ tests/utils/integration_test_case.py
@@ -1,5 +1,4 @@
 import os
-import platform
 import re
 import signal
 import unittest
